Final push

+++ src/pages/MyService.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../Provider/AuthProvieder';
 import { Link } from 'react-router';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const MyService = () => {
     const [myServices, setMyServices] = useState([]);
@@ -18,18 +19,46 @@ const MyService = () => {
     console.log(myServices);
 
     const handleDelete = (id) => {
-        axios.delete(`http://localhost:3000/delete/${id}`)
-            .then(res => {
-                console.log(res.data);
-                const filterData = myServices.filter(service => service._id != id)
-                console.log(filterData);
-                setMyServices(filterData)
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`http://localhost:3000/delete/${id}`)
+                    .then(res => {
+                        console.log(res.data);
+                        if (res.data.deletedCount == 1) {
 
-            })
-            .catch(err => {
-                console.log(err);
 
-            })
+                            const filterData = myServices.filter(service => service._id != id)
+                            console.log(filterData);
+                            setMyServices(filterData)
+
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            });
+                        }
+                    })
+                    .catch(err => {
+                        console.log(err);
+
+                    })
+
+
+
+
+
+            }
+        });
+
+
     }
     return (
         <div>
